Fixed step-in stopping in the wrong line in PPC64
Summary:
LLDB was not handling properly PPC64 functions that were
entered using the local entrypoint, instead of the global one.

The fix consists in comparing the current address after a function
call with both its global entrypoint as well as with the local entrypoint.

Also fixed a segmentation fault in PPC64 ABI when restarting a process.

Subscribers: nemanjai, kbarton

Differential Revision: https://reviews.llvm.org/D42582