Avoid GEP when creating a breakpoint
On PPC64le architecture, the function has extra instructions, which is the
global entry point (GEP), that adjusts the TOC pointer. When called locally,
this code is not executed, so when creating a breakpoint with the function
name, these instructions must be avoided, using the fist instruction after
the GEP section.